positive-reputation-py/requirements.txt
forta_agent>=0.1.5
setuptools>=61.3.1
pandas>=1.4.2
requests>=2.27.1
python-dotenv>=0.21.0